Factor Sector out into its own header. Sketch out Geometry.
